requirements.txt
fastapi~=0.128.0
python-dotenv~=1.0.1
python-telegram-bot~=22.7
md2tgmd~=0.1.2
pytz~=2024.1
pillow~=10.3.0
dateparser~=1.2.0
sqlalchemy[asyncio]~=2.0.46
alembic~=1.18.3
psycopg2-binary~=2.9.10
asyncpg~=0.31.0
aiosqlite~=0.22.1
google-genai~=1.61.0
pytest~=9.0.2
pytest-asyncio~=1.3.0
httpx~=0.28.1
pydantic~=2.12.5
mem0ai
anyio